+++ src/chunked.rs
@@ -25,15 +25,18 @@
 * Last updated: 2023-11-21
 */
 
-use crate::slicers::FnFindLastLineBreak;
-use arrow::array::ArrayRef;
+use arrow_array::array::ArrayRef;
 use parquet::format;
 use std::cmp::min;
+use std::fs;
+use crate::chunked::chunked_slicer::SLICER_IN_CHUNK_SIZE;
+
 
 // pub(crate) mod arrow2_converter;
 pub(crate) mod self_converter;
 
 pub mod arrow_converter;
+pub mod chunked_slicer;
 
 pub(crate) trait Converter<'a> {
     fn set_line_break_handler(&mut self, fn_line_break: FnFindLastLineBreak<'a>);
@@ -159,5 +162,111 @@ mod tests {
     #[test]
     fn test_column_length_num_rightaligned() {
         assert_eq!(column_length_num_rightaligned(b"123", 3), (0, 2));
+    }
+}
+
+
+pub(crate) struct ChunkAndResidue {
+    pub(crate) chunk: Box<[u8; SLICER_IN_CHUNK_SIZE]>,
+}
+pub(crate) trait Slicer<'a> {
+    fn slice_and_convert(
+        &mut self,
+        converter: Box<dyn 'a + Converter<'a>>,
+        infile: fs::File,
+        n_threads: usize,
+    ) -> Result<crate::chunked::Stats, &str>;
+}
+pub(crate) struct Stats {
+    pub(crate) bytes_in: usize,
+    pub(crate) bytes_out: usize,
+
+    pub(crate) num_rows: i64,
+}
+
+pub(crate) type FnLineBreakLen = fn() -> usize;
+#[allow(dead_code)]
+pub(crate) fn line_break_len_cr() -> usize {
+    1 as usize
+}
+#[allow(dead_code)]
+pub(crate) fn line_break_len_crlf() -> usize {
+    2 as usize
+}
+
+pub(crate) type FnFindLastLineBreak<'a> = fn(bytes: &'a [u8]) -> (bool, usize);
+#[allow(dead_code)]
+pub(crate) fn find_last_nlcr(bytes: &[u8]) -> (bool, usize) {
+    if bytes.is_empty() {
+        return (false, 0); // TODO should report err ...
+    }
+
+    let mut p2 = bytes.len() - 1;
+
+    if 0 == p2 {
+        return (false, 0); // hmm
+    }
+
+    loop {
+        if bytes[p2 - 1] == 0x0d && bytes[p2] == 0x0a {
+            return (true, p2 + 1);
+        }
+        if 0 == p2 {
+            return (false, 0); // indicate we didnt find nl
+        }
+
+        p2 -= 1;
+    }
+}
+
+#[allow(dead_code)]
+pub(crate) fn find_last_nl(bytes: &[u8]) -> (bool, usize) {
+    if bytes.is_empty() {
+        return (false, 0); // Indicate we didnt found nl.
+    }
+
+    let mut p2 = bytes.len() - 1;
+
+    if 0 == p2 {
+        return (false, 0); // hmm
+    }
+
+    loop {
+        if bytes[p2] == 0x0a {
+            return (true, p2);
+        }
+        if 0 == p2 {
+            return (false, 0); // indicate we didnt find nl
+        }
+        p2 -= 1;
+    }
+}
+pub(crate) struct IterRevolver<'a, T> {
+    shards: *mut T,
+    next: usize,
+    len: usize,
+    phantom: std::marker::PhantomData<&'a mut [T]>,
+}
+
+impl<'a, T> From<&'a mut [T]> for crate::chunked::IterRevolver<'a, T> {
+    fn from(shards: &'a mut [T]) -> crate::chunked::IterRevolver<'a, T> {
+        IterRevolver {
+            next: 0,
+            len: shards.len(),
+            shards: shards.as_mut_ptr(),
+            phantom: std::marker::PhantomData,
+        }
+    }
+}
+
+impl<'a, T> Iterator for IterRevolver<'a, T> {
+    type Item = &'a mut T;
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.next < self.len {
+            self.next += 1;
+        } else {
+            self.next = 1;
+        }
+        unsafe { Some(&mut *self.shards.offset(self.next as isize - 1)) }
     }
 }
