+++ src/chunked/arrow_converter.rs
@@ -40,9 +40,9 @@ use rayon::iter::IndexedParallelIterator;
 use rayon::prelude::*;
 
 use crate::datatype::DataType;
-use crate::converters::{ColumnBuilder, Converter};
-use crate::slicers::{FnFindLastLineBreak, FnLineBreakLen};
-use crate::{converters, schema};
+use crate::chunked::{ColumnBuilder, Converter};
+use crate::chunked::{FnFindLastLineBreak, FnLineBreakLen};
+use crate::{chunked, schema};
 use debug_print::debug_println;
 
 pub(crate) struct Slice2Arrow<'a> {
@@ -224,7 +224,7 @@ impl ColumnBuilder for HandlerInt32Builder {
         Self: Sized,
     {
         let (start, stop) =
-            converters::column_length_num_rightaligned(data, self.runes_in_column as i16);
+            chunked::column_length_num_rightaligned(data, self.runes_in_column as i16);
 
         match atoi_simd::parse(&data[start..stop]) {
             Ok(n) => {
@@ -255,7 +255,7 @@ impl ColumnBuilder for HandlerInt64Builder {
         Self: Sized,
     {
         let (start, stop) =
-            converters::column_length_num_rightaligned(data, self.runes_in_column as i16);
+            chunked::column_length_num_rightaligned(data, self.runes_in_column as i16);
         match atoi_simd::parse(&data[start..stop]) {
             Ok(n) => {
                 self.int64builder.append_value(n);
@@ -287,7 +287,7 @@ impl ColumnBuilder for HandlerStringBuilder {
     where
         Self: Sized,
     {
-        let column_length: usize = converters::column_length(data, self.runes_in_column as i16);
+        let column_length: usize = chunked::column_length(data, self.runes_in_column as i16);
         // Me dont like ... what is the cost ? Could it be done once for the whole chunk ?
         let text: &str = unsafe { from_utf8_unchecked(&data[..column_length]) };
 
@@ -323,7 +323,7 @@ impl ColumnBuilder for HandlerBooleanBuilder {
         Self: Sized,
     {
         let (start, stop) =
-            converters::column_length_char_rightaligned(data, self.runes_in_column as i16);
+            chunked::column_length_char_rightaligned(data, self.runes_in_column as i16);
 
         let text: &str = unsafe { from_utf8_unchecked(&data[start..stop]) };
