+++ src/cli.rs
@@ -29,14 +29,14 @@ use std::fs;
 use std::fs::File;
 use std::path::PathBuf;
 
-// use crate::converters::arrow2_converter::{MasterBuilder, Slice2Arrow2};
-use crate::converters::arrow_converter::{MasterBuilders, Slice2Arrow};
-use crate::converters::self_converter::SampleSliceAggregator;
-use crate::converters::Converter;
+// use crate::chunked::arrow2_converter::{MasterBuilder, Slice2Arrow2};
+use crate::chunked::arrow_converter::{MasterBuilders, Slice2Arrow};
+use crate::chunked::self_converter::SampleSliceAggregator;
+use crate::chunked::Converter;
 use crate::dump::dump;
-use crate::slicers::chunked_slicer::{OldSlicer, IN_MAX_CHUNKS, SLICER_IN_CHUNK_SIZE};
-use crate::slicers::Slicer;
-use crate::slicers::{find_last_nl, line_break_len_cr, ChunkAndResidue};
+use crate::chunked::chunked_slicer::{OldSlicer, IN_MAX_CHUNKS, SLICER_IN_CHUNK_SIZE};
+use crate::chunked::Slicer;
+use crate::chunked::{find_last_nl, line_break_len_cr, ChunkAndResidue};
 use crate::{converter, error, mocker, schema};
 use clap::{value_parser, ArgAction, Parser, Subcommand};
 use log::info;
